+++ apps/cowswap-frontend/src/modules/trade/pure/Settings/index.tsx
@@ -6,8 +6,6 @@ export {
   SettingsBoxTitle,
   SettingsBoxWrapper,
   SettingsButton,
-  SettingsButtonIcon,
-  UsdButton,
 } from './styled'
 export { SettingsBox } from './SettingsBox'
 export { SettingsIcon } from './SettingsIcon'