+++ apps/cowswap-frontend/src/modules/yield/hooks/useYieldDerivedState.ts
@@ -1,30 +1,24 @@
 import { useAtomValue } from 'jotai'
-import { useSetAtom } from 'jotai'
-import { useEffect } from 'react'
+import { useMemo } from 'react'
 
 import { INITIAL_ALLOWED_SLIPPAGE_PERCENT } from '@cowprotocol/common-const'
 
 import { TradeType, useBuildTradeDerivedState } from 'modules/trade'
 
-import { yieldDerivedStateAtom, yieldRawStateAtom } from '../state/yieldRawStateAtom'
+import { YieldDerivedState, yieldDerivedStateAtom, yieldRawStateAtom } from '../state/yieldRawStateAtom'
 
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export function useYieldDerivedState() {
+export function useYieldDerivedState(): YieldDerivedState {
   return useAtomValue(yieldDerivedStateAtom)
 }
 
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export function useFillYieldDerivedState() {
-  const updateDerivedState = useSetAtom(yieldDerivedStateAtom)
+export function useYieldDerivedStateToFill(): YieldDerivedState {
   const derivedState = useBuildTradeDerivedState(yieldRawStateAtom, false)
 
-  useEffect(() => {
-    updateDerivedState({
+  return useMemo(() => {
+    return {
       ...derivedState,
       slippage: INITIAL_ALLOWED_SLIPPAGE_PERCENT,
       tradeType: TradeType.YIELD,
-    })
-  }, [derivedState, updateDerivedState])
+    }
+  }, [derivedState])
 }